update for test case 3

+++ C++ For DSA/Module-2.5/problem-2.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int X, Y;
+    long long X, Y;
     cin >> X >> Y;
     cout << X << " + " << Y << " = " << X + Y << endl;
     cout << X << " * " << Y << " = " << X * Y << endl;
